Add multi-threading support to secmod module.
The secmod binary handles crypto operations and authentications. Since the
binary was mono-threaded and given how requests from workers were handled, it
was not possible to have multiple authentications in parallel or even crypto
operations and authentications in parallel. For example, a request triggering
long blocking IO could degrade performance by blocking authentications and
operations.

Some identity providers provide their own PAM modules as a generic way to
handle authentications and some of them are implemented with blocking IO making
a call to pam_authenticate blocking the main and only thread.
For instance, the Duo PAM module used for push notifications posts an HTTP
query and wait for the answer. Since the request handling is mono-threaded,
this creates contention as no other user can authenticate and no operation
could complete until the user has either acknowledged the request or the time
out is reached.

Adding multi-threading support to secmod allows to handle as many crypto
operations and authentications in parallel as there are threads.
By default the number of threads is equal to the number of cores available on
the machine but this will be made configurable in a future commit. Since
multi-threading has been introduced, some data structures had to be protected
againt concurrent accesses.

An MIT-licensed thread pool implementation has been chosen to be included to
the project in order to avoid creating a new dependency. The original project
is available at https://github.com/Pithikos/C-Thread-Pool.

Signed-off-by: Clement Michaud <clement.michaud34@gmail.com>