Protect secmod module against DoS attacks.
Now that secmod is multi-threaded an attacker could exhaust memory by sending
lots of authentication requests that will eventually be converted into tasks
stored in the linked list representing the tasks to be treated by the thread
pool.

In order to mitigate and improve time to recovery when facing an attack where
too many requests are received, two options are introduced to limit the
accumulation of tasks in the queue and therefore prevent memory exhaustion in
secmod.

The options are:

* secmod-max-requests (uint) defines the max number of requests secmod can
queue in memory. Default: 0, meaning no limitation.

* secmod-drop-requests (bool) declares whether requests should be dropped or
throttled. Default: false, meaning some requests are throttled when more than
secmod-max-requests requests have been sent.

Signed-off-by: Clement Michaud <clement.michaud34@gmail.com>